+++ src/commands/check.ts
@@ -2,6 +2,7 @@ import { Project } from '../common/Config';
 import { PipelineCommand } from '../common/pipelines';
 import { Response, StatusCode } from '../common/api';
 import { IPipeline } from '../common/iPipeline';
+import { bold } from 'chalk';
 
 export class Check extends PipelineCommand {
   protected async runPerProject(project: Project) {
@@ -16,8 +17,12 @@ export class Check extends PipelineCommand {
   private async checkStatus(project: Project, pipeline: Response<IPipeline>) {
     const data = pipeline.data;
     if (pipeline.status === StatusCode.Success) {
-      const msg = `Pipeline status: ${data.status} last update ${data.created_at}`;
-      this.screenPrinter.setProjectSuccess(project, msg);
+      const msg = `Pipeline status: ${bold(data.status)} last update ${data.created_at}`;
+      if (data.status === 'success') {
+        this.screenPrinter.setProjectSuccess(project, msg);
+      } else {
+        this.screenPrinter.setProjectWarn(project, msg);
+      }
     }
     return pipeline;
   }